+++ SmartDeviceLink/SDLDeleteCommand.h
@@ -32,7 +32,7 @@
  */
 - (instancetype)initWithDictionary:(NSMutableDictionary *)dict;
 
-- (instancetype)initWithId:(NSInteger)commandId;
+- (instancetype)initWithId:(UInt32)commandId;
 
 /**
  * @abstract the Command ID that identifies the Command to be deleted from Command Menu
