+++ SmartDeviceLink/SDLMenuParams.h
@@ -25,7 +25,7 @@
 
 - (instancetype)initWithMenuName:(NSString*)menuName;
 
-- (instancetype)initWithMenuName:(NSString *)menuName parentId:(NSUInteger)parentId position:(NSUInteger)position;
+- (instancetype)initWithMenuName:(NSString *)menuName parentId:(UInt32)parentId position:(UInt16)position;
 
 /**
  * @abstract the unique ID of an existing submenu to which a command will be added
