+++ SmartDeviceLink/SDLStartTime.m
@@ -20,7 +20,7 @@
     return self;
 }
 
-- (instancetype)initWithHours:(NSInteger)hours minutes:(NSInteger)minutes seconds:(NSInteger)seconds {
+- (instancetype)initWithHours:(UInt8)hours minutes:(UInt8)minutes seconds:(UInt8)seconds {
     self = [self init];
     if (!self) {
         return nil;
